+++ src/pages/UserInfoPage/components/DeleteAccountButton.tsx
@@ -4,7 +4,7 @@ const DeleteAccountButton = () => {
       type='button'
       className='h-[23px] w-[100%] text-[12px] text-subfont underline active:text-black'
     >
-      회원 탈퇴하기
+      탈퇴하시겠습니까?
     </button>
   );
 };
