+++ src/pages/UserInfoPage/components/InfoForm.tsx
@@ -1,11 +1,11 @@
 import DeleteAccountButton from './DeleteAccountButton';
-import InputContainer from './InputContainer';
+import InfoContainer from './InfoContainer';
 
 const InfoForm = () => {
   return (
-    <div className='mt-14 flex w-[375px] flex-col justify-center'>
-      <div className='mx-auto flex w-custom flex-col justify-center gap-10'>
-        <InputContainer />
+    <div className='mt-14 flex w-[375px] flex-col justify-center gap-48'>
+      <div className='mx-auto flex w-custom flex-col justify-center gap-14'>
+        <InfoContainer />
         <button
           type='button'
           className='btn-primary border border-primary bg-white text-primary active:bg-primary active:text-white'
@@ -13,9 +13,7 @@ const InfoForm = () => {
           수정하기
         </button>
       </div>
-      <div>
-        <DeleteAccountButton />
-      </div>
+      <DeleteAccountButton />
     </div>
   );
 };
